+++ server/src/tests/cases/compiler/array6.c
@@ -1,0 +1,8 @@
+* test() {
+
+    string *arr = ( { "a", "b" } );
+    return arr;
+    
+}
+
+// declaration can have an array indicator without a type